+++ crates/monochange_cargo/Cargo.toml
@@ -16,6 +16,7 @@ description = "Manage versions and releases for your multiplatform, multilanguag
 glob = { workspace = true, default-features = true }
 monochange_core = { workspace = true }
 monochange_linting = { workspace = true }
+monochange_publish = { workspace = true }
 monochange_semver = { workspace = true }
 quote = { workspace = true, default-features = true }
 semver = { workspace = true, default-features = true }
